requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.1
pandas==2.1.4
numpy==1.26.3
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.0
Pillow==10.2.0
